feat: RssCategoryfield as component
