+++ packages/app/src/views/edit-feed/itunes.tsx
@@ -2,20 +2,17 @@ import { Container } from "@fourviere/ui/lib/box";
 import FormSection from "@fourviere/ui/lib/form/form-section";
 import FormRow from "@fourviere/ui/lib/form/form-row";
 import Input from "@fourviere/ui/lib/form/fields/input";
-import { Field, FieldArray, Formik } from "formik";
+import { Formik } from "formik";
 import { FormField } from "@fourviere/ui/lib/form/form-field";
 import UseCurrentFeed from "../../hooks/useCurrentFeed";
 import useTranslations from "../../hooks/useTranslations";
-import Undefined from "@fourviere/ui/lib/form/fields/undefined";
-import Select from "@fourviere/ui/lib/form/fields/select";
 import FormObserver from "../../components/form-observer";
 import { Feed } from "@fourviere/core/lib/schema/feed";
-import { APPLE_PODCAST_CATEGORIES } from "../../consts";
-import ResetField from "@fourviere/ui/lib/form/reset-field";
 import FormObjectField from "@fourviere/ui/lib/form/form-object-field";
 import Boolean from "@fourviere/ui/lib/form/fields/boolean";
 import ImageField from "@fourviere/ui/lib/form/fields/image";
 import useUpload from "../../hooks/useUpload";
+import {RssChannelItunesCategoryFields} from "../../components/form-fields/RssChannelItunesCategoryFields.tsx";
 export default function Itunes() {
   const currentFeed = UseCurrentFeed();
   const t = useTranslations();
@@ -34,6 +31,7 @@ export default function Itunes() {
       }}
     >
       {({ values, handleSubmit, setFieldValue, setFieldError }) => {
+        console.log({values})
         const imageUpload = useUpload({
           feedId: currentFeed.feedId,
           updateField: (value: string) =>
@@ -138,79 +136,7 @@ export default function Itunes() {
                   emtpyValueButtonMessage={t["ui.forms.empty_field.message"]}
                 />
               </FormRow>
-
-              <FieldArray name={`rss.channel.0.["itunes:category"]`}>
-                {({ remove, push }) => (
-                  <FormRow
-                    name="itunes_ext.categories"
-                    label="itunes categories"
-                  >
-                    <div className="space-y-1">
-                      {values.rss.channel[0]["itunes:category"].map(
-                        (category, index) => {
-                          const subCategories = APPLE_PODCAST_CATEGORIES.find(
-                            (e) => e.category === category["@"].text
-                          )?.subcategories.map((e) => ({
-                            key: e,
-                            value: e,
-                          }));
-
-                          return (
-                            <Container
-                              spaceX="sm"
-                              flex="row-middle"
-                              key={index}
-                            >
-                              <Field
-                                as={Select}
-                                keyProperty="category"
-                                labelProperty="category"
-                                name={`rss.channel.0.['itunes:category'].${index}.@.text`}
-                                options={APPLE_PODCAST_CATEGORIES.map((e) => ({
-                                  category: e.category,
-                                }))}
-                              />
-
-                              {!!subCategories?.length && (
-                                <Field
-                                  as={Select}
-                                  keyProperty="key"
-                                  labelProperty="value"
-                                  name={`rss.channel.0.['itunes:category'].${index}.['itunes:category'].@.text`}
-                                  options={subCategories}
-                                />
-                              )}
-                              <div>
-                                <ResetField
-                                  onClick={() => {
-                                    remove(index);
-                                  }}
-                                />
-                              </div>
-                            </Container>
-                          );
-                        }
-                      )}
-
-                      {values.rss.channel[0]["itunes:category"].length < 2 && (
-                        <Undefined
-                          onClick={() =>
-                            push({
-                              "itunes:category": {
-                                "@": { text: "Hobbies" },
-                              },
-                              "@": { text: "Leisure" },
-                            })
-                          }
-                        >
-                          {t["ui.forms.empty_field.message"]}
-                        </Undefined>
-                      )}
-                    </div>
-                  </FormRow>
-                )}
-              </FieldArray>
-
+              <RssChannelItunesCategoryFields name={`rss.channel.0.["itunes:category"]`} />
               <FormRow
                 name={`rss.channel.0.["itunes:complete"]`}
                 label={t["edit_feed.channel_field.itunes.complete"]}
